CU-8678e8320 - ClickClock: criar modal de login que usa API nova do ClickClock
